rag_pdf_platform/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
pydantic-settings>=2.1.0
pypdf2>=3.0.1
chromadb>=0.4.24
sentence-transformers>=2.2.2
openai>=1.40.0
